Refactor the single main.go file into three distinct files to improve code structure and reusability.
This change separates the core application logic into:
- `router.go`: Contains the generic message routing logic.
- `consumer.go`: Contains the SQS consumer logic for polling and processing messages.
- `main.go`: Acts as the application entry point, responsible for initializing and connecting the router and consumer components.

This refactoring makes the router and consumer components more modular and reusable, in preparation for potentially moving the main application logic to an `examples/` directory in the future.
