+++ .gitignore
@@ -27,6 +27,9 @@ go.work.sum
 # env file
 .env
 
+# Build output
+sqsrouter
+
 # Editor/IDE
 # .idea/
 # .vscode/
